refactor: improve file naming conventions
